print elements

+++ 0x17-doubly_linked_lists/1-dlistint_len.c
@@ -8,7 +8,7 @@
 
 size_t dlistint_len(const dlistint_t *h)
 {
-	int l = 0;
+	size_t l = 0;
 
 	if (h == NULL)
 	{
